requirements.txt
scapy
pefile
python-whois
WMI
requests